resolutions 4; unmerged local branch changes of inheritance to ZoneAware, AmenityOwner, etc.
